+++ ksql-engine/src/test/java/io/confluent/ksql/ddl/commands/CreateSourceCommandTest.java
@@ -336,6 +336,29 @@ public class CreateSourceCommandTest {
     )));
   }
 
+  @Test
+  public void shouldValidateValueFormatCanHandleValueSchema() {
+    // Given:
+    final LogicalSchema schema = LogicalSchema.of(SchemaBuilder
+        .struct()
+        .field("bob", Schema.OPTIONAL_STRING_SCHEMA)
+        .field("hojjat", Schema.OPTIONAL_STRING_SCHEMA)
+        .build());
+
+    // When:
+    new TestCmd(
+        "sql",
+        statement,
+        ksqlConfig,
+        kafkaTopicClient,
+        serdeOptions,
+        serdeFactories
+    );
+
+    // Then:
+    verify(serdeFactory).validate(schema.valueSchema());
+  }
+
   private static Map<String, Literal> minValidProps() {
     return ImmutableMap.of(
         CommonCreateConfigs.VALUE_FORMAT_PROPERTY, new StringLiteral("json"),