+++ ksql-parser/src/main/java/io/confluent/ksql/parser/tree/TableElements.java
@@ -17,10 +17,14 @@ package io.confluent.ksql.parser.tree;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Iterables;
 import com.google.errorprone.annotations.Immutable;
 import io.confluent.ksql.parser.tree.TableElement.Namespace;
+import io.confluent.ksql.schema.ksql.LogicalSchema;
+import io.confluent.ksql.schema.ksql.SchemaConverters;
 import io.confluent.ksql.schema.ksql.SqlBaseType;
 import io.confluent.ksql.util.KsqlException;
+import io.confluent.ksql.util.SchemaUtil;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Iterator;
@@ -31,6 +35,8 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
+import org.apache.kafka.connect.data.Schema;
+import org.apache.kafka.connect.data.SchemaBuilder;
 
 @Immutable
 public final class TableElements implements Iterable<TableElement> {
@@ -74,6 +80,32 @@ public final class TableElements implements Iterable<TableElement> {
   @Override
   public String toString() {
     return elements.toString();
+  }
+
+  public LogicalSchema toLogicalSchema() {
+    if (Iterables.isEmpty(this)) {
+      throw new KsqlException("No columns supplied.");
+    }
+
+    final SchemaBuilder keySchema = SchemaBuilder.struct();
+    final SchemaBuilder valueSchema = SchemaBuilder.struct();
+    for (final TableElement tableElement : this) {
+      final String fieldName = tableElement.getName();
+      final Schema fieldSchema = SchemaConverters.sqlToLogicalConverter()
+          .fromSqlType(tableElement.getType().getSqlType());
+
+      if (tableElement.getNamespace() == Namespace.KEY) {
+        keySchema.field(fieldName, fieldSchema);
+      } else {
+        valueSchema.field(fieldName, fieldSchema);
+      }
+    }
+
+    if (keySchema.fields().isEmpty()) {
+      keySchema.field(SchemaUtil.ROWKEY_NAME, Schema.OPTIONAL_STRING_SCHEMA);
+    }
+
+    return LogicalSchema.of(keySchema.build(), valueSchema.build());
   }
 
   private TableElements(final ImmutableList<TableElement> elements) {