+++ ksql-serde/src/test/java/io/confluent/ksql/serde/KsqlSerdeFactoryTest.java
@@ -137,6 +137,10 @@ public class KsqlSerdeFactoryTest {
     }
 
     @Override
+    public void validate(final ConnectSchema schema) {
+    }
+
+    @Override
     protected Serializer<Object> createSerializer(
         final PersistenceSchema schema,
         final KsqlConfig ksqlConfig,